ventas

+++ database/migrations/2025_03_29_000003_create_ventas_table.php
@@ -7,15 +7,18 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('ventas', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('cliente_id')->constrained('clientes')->onDelete('cascade');
-            $table->foreignId('id_empleado')->constrained('empleados')->onDelete('cascade');
-            $table->date('fecha');
-            $table->decimal('total', 10, 2);
-            $table->enum('estado', ['Pendiente', 'Pagada', 'Cancelada'])->default('Pendiente');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('ventas')) { // Verificar si la tabla ya existe
+            Schema::create('ventas', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('cliente_id')->constrained('clientes')->onDelete('cascade');
+                $table->foreignId('id_empleado')->constrained('empleados')->onDelete('cascade');
+                $table->date('fecha');
+                $table->decimal('total', 10, 2);
+                $table->string('rtn')->nullable();
+                $table->enum('estado', ['Pendiente', 'Pagada', 'Cancelada'])->default('Pendiente');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
